agent/requirements.txt
livekit-agents>=1.0
livekit-plugins-aws[realtime]
livekit-plugins-silero
livekit-plugins-turn-detector
livekit-api
python-dotenv
aiohttp
boto3
